CLOUD/gcp/requirements.txt
flask==3.1.0
google-cloud-vision==3.9.0